Adding Test project and making it build
