--- DateTimeExtensions.Tests/DateTimeExtensions.Tests.csproj
@@ -1,95 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProductVersion>8.0.30703</ProductVersion>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{09E6DAA1-36B6-4835-8379-7D819AB33066}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>DateTimeExtensions.Tests</RootNamespace>
-    <AssemblyName>DateTimeExtensions.Tests</AssemblyName>
-    <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <SolutionDir Condition="$(SolutionDir) == '' Or $(SolutionDir) == '*Undefined*'">..\</SolutionDir>
-    <RestorePackages>true</RestorePackages>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="NSubstitute">
-      <HintPath>..\packages\NSubstitute.1.3.0.0\lib\NET35\NSubstitute.dll</HintPath>
-    </Reference>
-    <Reference Include="nunit.framework, Version=2.6.0.12051, Culture=neutral, PublicKeyToken=96d09a1eb7f44a77, processorArchitecture=MSIL">
-      <SpecificVersion>False</SpecificVersion>
-      <HintPath>..\packages\NUnit.2.6.0.12054\lib\nunit.framework.dll</HintPath>
-    </Reference>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="ARSAWorkingDays.cs" />
-    <Compile Include="ChineseHolidaysTests.cs" />
-    <Compile Include="ChristianHolidaysTests.cs" />
-    <Compile Include="HolidaysTranslations\GermanHolidayNamesTest.cs" />
-    <Compile Include="HolidaysTranslations\ViHolidaysNamesTest.cs" />
-    <Compile Include="ViHolidaysTests.cs" />
-    <Compile Include="HolidaysTranslations\KRHolidaysNamesTest.cs" />
-    <Compile Include="KRHolidaysTests.cs" />
-    <Compile Include="KRNaturalTimeTests.cs" />
-    <Compile Include="en-NZCalendarTest.cs" />
-    <Compile Include="HolidaysTranslations\DefaultNamesTest.cs" />
-    <Compile Include="HolidaysTranslations\SpanishHolidaysNamesTest.cs" />
-    <Compile Include="NLHolidaysTests.cs" />
-    <Compile Include="NLNaturalTimeTests.cs" />
-    <Compile Include="FrCaHolidaysTests.cs" />
-    <Compile Include="GenericNaturalTimeTests.cs" />
-    <Compile Include="ExampleTests.cs" />
-    <Compile Include="GenericCalendarTests.cs" />
-    <Compile Include="FRNaturalTimeTests.cs" />
-    <Compile Include="HolidaysTest.cs" />
-    <Compile Include="JewishHolidaysTests.cs" />
-    <Compile Include="OfficeHolidayExportHolidaysFormatTests.cs" />
-    <Compile Include="PTCalendarTests.cs" />
-    <Compile Include="HolidaysTranslations\PortugueseHolidaysNamesTest.cs" />
-    <Compile Include="PTWorkingDayCultureInfoTests.cs" />
-    <Compile Include="GenericWorkingDayCultureInfoTests.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RoundCeilingFloorDatesTests.cs" />
-    <Compile Include="TimeTests.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
-       Other similar extension points exist, see Microsoft.Common.targets.
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-  -->
-</Project>
