+++ DateTimeExtensions.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DateTimeExtensions.Tests", "DateTimeExtensions.Tests\DateTimeExtensions.Tests.csproj", "{09E6DAA1-36B6-4835-8379-7D819AB33066}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{ED60F95F-4CCC-4F56-ACB2-DEE39CB00A54}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
@@ -19,25 +17,30 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{EB805547-455
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "DateTimeExtensions", "src\DateTimeExtensions\DateTimeExtensions.xproj", "{650FA9EC-03BE-43AF-87FF-9632856E31EE}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{0F8CA42F-760A-4A42-BD0F-0AC9900B6712}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "DateTimeExtensions.Tests", "tests\DateTimeExtensions.Tests\DateTimeExtensions.Tests.xproj", "{2BB0401A-4DCA-4D59-B15D-6C895304D2AF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{09E6DAA1-36B6-4835-8379-7D819AB33066}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{09E6DAA1-36B6-4835-8379-7D819AB33066}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{09E6DAA1-36B6-4835-8379-7D819AB33066}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{09E6DAA1-36B6-4835-8379-7D819AB33066}.Release|Any CPU.Build.0 = Release|Any CPU
 		{650FA9EC-03BE-43AF-87FF-9632856E31EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{650FA9EC-03BE-43AF-87FF-9632856E31EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{650FA9EC-03BE-43AF-87FF-9632856E31EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{650FA9EC-03BE-43AF-87FF-9632856E31EE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2BB0401A-4DCA-4D59-B15D-6C895304D2AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2BB0401A-4DCA-4D59-B15D-6C895304D2AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2BB0401A-4DCA-4D59-B15D-6C895304D2AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2BB0401A-4DCA-4D59-B15D-6C895304D2AF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{650FA9EC-03BE-43AF-87FF-9632856E31EE} = {EB805547-4556-4744-AFF4-314BE0F45704}
+		{2BB0401A-4DCA-4D59-B15D-6C895304D2AF} = {0F8CA42F-760A-4A42-BD0F-0AC9900B6712}
 	EndGlobalSection
 EndGlobal
