+++ tests/DateTimeExtensions.Tests/FRNaturalTimeTests.cs
@@ -20,7 +20,8 @@ namespace DateTimeExtensions.Tests
 
             var naturalText = fromTime.ToNaturalText(toTime, false, foo_ci);
 
-            Assert.IsNotNullOrEmpty(naturalText);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
             Assert.AreEqual("2 heures", naturalText);
         }
 
@@ -32,7 +33,8 @@ namespace DateTimeExtensions.Tests
 
             var naturalText = fromTime.ToNaturalText(toTime, true, foo_ci);
 
-            Assert.IsNotNullOrEmpty(naturalText);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
             Assert.AreEqual("3 heures", naturalText);
         }
 
@@ -45,7 +47,8 @@ namespace DateTimeExtensions.Tests
 
             var naturalText = fromTime.ToExactNaturalText(toTime, foo_ci);
 
-            Assert.IsNotNullOrEmpty(naturalText);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
             Assert.AreEqual("2 heures, 30 minutes", naturalText);
         }
 
@@ -57,7 +60,8 @@ namespace DateTimeExtensions.Tests
 
             var naturalText = fromTime.ToExactNaturalText(toTime, foo_ci);
 
-            Assert.IsNotNullOrEmpty(naturalText);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
             Assert.AreEqual("2 ans, 2 mois, 3 journées, 4 heures, 5 minutes, 6 seconds", naturalText);
         }
 
@@ -70,8 +74,10 @@ namespace DateTimeExtensions.Tests
             var naturalText = fromTime.ToExactNaturalText(toTime, foo_ci);
             var naturalTextReverse = toTime.ToExactNaturalText(fromTime, foo_ci);
 
-            Assert.IsNotNullOrEmpty(naturalText);
-            Assert.IsNotNullOrEmpty(naturalTextReverse);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
+            Assert.IsNotNull(naturalTextReverse);
+            Assert.IsNotEmpty(naturalTextReverse);
             Assert.AreEqual(naturalTextReverse, naturalText);
         }
 
@@ -86,8 +92,10 @@ namespace DateTimeExtensions.Tests
             var naturalText_plural = fromTime.ToNaturalText(toTime_plural, true, foo_ci);
             var naturalText_single = fromTime.ToNaturalText(toTime_single, true, foo_ci);
 
-            Assert.IsNotNullOrEmpty(naturalText_plural);
-            Assert.IsNotNullOrEmpty(naturalText_single);
+            Assert.IsNotNull(naturalText_plural);
+            Assert.IsNotEmpty(naturalText_plural);
+            Assert.IsNotNull(naturalText_single);
+            Assert.IsNotEmpty(naturalText_single);
             Assert.AreEqual("2 ans", naturalText_plural);
             Assert.AreEqual("1 année", naturalText_single);
         }
@@ -102,8 +110,10 @@ namespace DateTimeExtensions.Tests
             var naturalText_plural = fromTime.ToNaturalText(toTime_plural, true, foo_ci);
             var naturalText_single = fromTime.ToNaturalText(toTime_single, true, foo_ci);
 
-            Assert.IsNotNullOrEmpty(naturalText_plural);
-            Assert.IsNotNullOrEmpty(naturalText_single);
+            Assert.IsNotNull(naturalText_plural);
+            Assert.IsNotEmpty(naturalText_plural);
+            Assert.IsNotNull(naturalText_single);
+            Assert.IsNotEmpty(naturalText_single);
             Assert.AreEqual("2 mois", naturalText_plural);
             Assert.AreEqual("1 mois", naturalText_single);
         }
@@ -118,8 +128,10 @@ namespace DateTimeExtensions.Tests
             var naturalText_plural = fromTime.ToExactNaturalText(toTime_plural, foo_ci);
             var naturalText_single = fromTime.ToExactNaturalText(toTime_single, foo_ci);
 
-            Assert.IsNotNullOrEmpty(naturalText_plural);
-            Assert.IsNotNullOrEmpty(naturalText_single);
+            Assert.IsNotNull(naturalText_plural);
+            Assert.IsNotEmpty(naturalText_plural);
+            Assert.IsNotNull(naturalText_single);
+            Assert.IsNotEmpty(naturalText_single);
             Assert.AreEqual("2 journées, 2 heures, 2 minutes, 2 seconds", naturalText_plural);
             Assert.AreEqual("1 jour, 1 heure, 1 minute, 1 second", naturalText_single);
         }
