+++ tests/DateTimeExtensions.Tests/HolidaysTranslations/PortugueseHolidaysNamesTest.cs
@@ -16,7 +16,7 @@ namespace DateTimeExtensions.Tests
         public void Setup()
         {
             //setup a default culture
-            System.Threading.Thread.CurrentThread.CurrentUICulture = new CultureInfo("en-US");
+            CultureInfo.CurrentUICulture = new CultureInfo("en-US");
         }
 
         [Test]
@@ -28,8 +28,8 @@ namespace DateTimeExtensions.Tests
             Assert.AreEqual(PT_PTHolidayStrategy.RepublicDay.Name, "Republic Day");
             Assert.AreEqual(PT_PTHolidayStrategy.RestorationOfIndependance.Name, "Restoration of Independance");
 
-            System.Threading.Thread.CurrentThread.CurrentUICulture = new CultureInfo("pt-PT");
-            Assert.AreEqual("pt-PT", System.Threading.Thread.CurrentThread.CurrentUICulture.Name);
+            CultureInfo.CurrentUICulture = new CultureInfo("pt-PT");
+            Assert.AreEqual("pt-PT", CultureInfo.CurrentUICulture.Name);
 
             Assert.AreEqual(PT_PTHolidayStrategy.FreedomDay.Name, "Dia da Liberdade");
             Assert.AreEqual(PT_PTHolidayStrategy.PortugalDay.Name, "Dia de Portugal");
@@ -51,7 +51,6 @@ namespace DateTimeExtensions.Tests
         }
 
         [Test]
-        [SetUICulture("pt-BR")]
         public void AssertBrazilianHolidaysAreTranslated()
         {
             //test holidays still on default culture (en-US)
@@ -62,7 +61,8 @@ namespace DateTimeExtensions.Tests
             Assert.AreEqual(PT_BRHolidayStrategy.TiradentesDay.Name, "Tiradentes Day");
             */
             //System.Threading.Thread.CurrentThread.CurrentUICulture = new CultureInfo("pt-BR");
-            Assert.AreEqual("pt-BR", System.Threading.Thread.CurrentThread.CurrentUICulture.Name);
+            CultureInfo.CurrentUICulture = new CultureInfo("pt-BR");
+            Assert.AreEqual("pt-BR", CultureInfo.CurrentUICulture.Name);
 
             Assert.AreEqual(PT_BRHolidayStrategy.IndependanceDay.Name, "Dia da Independência");
             Assert.AreEqual(PT_BRHolidayStrategy.OurLadyOfAparecida.Name, "Nossa Srª da Aparecida");
