+++ tests/DateTimeExtensions.Tests/NLNaturalTimeTests.cs
@@ -35,7 +35,8 @@ namespace DateTimeExtensions.Tests
 
             var naturalText = fromTime.ToNaturalText(toTime, false, this.culture);
 
-            Assert.IsNotNullOrEmpty(naturalText);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
             Assert.AreEqual("2 uur", naturalText);
         }
 
@@ -47,7 +48,8 @@ namespace DateTimeExtensions.Tests
 
             var naturalText = fromTime.ToNaturalText(toTime, true, this.culture);
 
-            Assert.IsNotNullOrEmpty(naturalText);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
             Assert.AreEqual("3 uur", naturalText);
         }
 
@@ -60,7 +62,8 @@ namespace DateTimeExtensions.Tests
 
             var naturalText = fromTime.ToExactNaturalText(toTime, this.culture);
 
-            Assert.IsNotNullOrEmpty(naturalText);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
             Assert.AreEqual("2 uur, 30 minuten", naturalText);
         }
 
@@ -72,7 +75,8 @@ namespace DateTimeExtensions.Tests
 
             var naturalText = fromTime.ToExactNaturalText(toTime, this.culture);
 
-            Assert.IsNotNullOrEmpty(naturalText);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
             Assert.AreEqual("2 jaar, 2 maanden, 3 dagen, 4 uur, 5 minuten, 6 seconden", naturalText);
         }
 
@@ -85,8 +89,10 @@ namespace DateTimeExtensions.Tests
             var naturalText = fromTime.ToExactNaturalText(toTime, this.culture);
             var naturalTextReverse = toTime.ToExactNaturalText(fromTime, this.culture);
 
-            Assert.IsNotNullOrEmpty(naturalText);
-            Assert.IsNotNullOrEmpty(naturalTextReverse);
+            Assert.IsNotNull(naturalText);
+            Assert.IsNotEmpty(naturalText);
+            Assert.IsNotNull(naturalTextReverse);
+            Assert.IsNotEmpty(naturalTextReverse);
             Assert.AreEqual(naturalTextReverse, naturalText);
         }
 
@@ -101,8 +107,10 @@ namespace DateTimeExtensions.Tests
             var naturalText_plural = fromTime.ToNaturalText(toTime_plural, true, this.culture);
             var naturalText_single = fromTime.ToNaturalText(toTime_single, true, this.culture);
 
-            Assert.IsNotNullOrEmpty(naturalText_plural);
-            Assert.IsNotNullOrEmpty(naturalText_single);
+            Assert.IsNotNull(naturalText_plural);
+            Assert.IsNotEmpty(naturalText_plural);
+            Assert.IsNotNull(naturalText_single);
+            Assert.IsNotEmpty(naturalText_single);
             Assert.AreEqual("2 jaar", naturalText_plural);
             Assert.AreEqual("1 jaar", naturalText_single);
         }
@@ -117,8 +125,10 @@ namespace DateTimeExtensions.Tests
             var naturalText_plural = fromTime.ToNaturalText(toTime_plural, true, this.culture);
             var naturalText_single = fromTime.ToNaturalText(toTime_single, true, this.culture);
 
-            Assert.IsNotNullOrEmpty(naturalText_plural);
-            Assert.IsNotNullOrEmpty(naturalText_single);
+            Assert.IsNotNull(naturalText_plural);
+            Assert.IsNotEmpty(naturalText_plural);
+            Assert.IsNotNull(naturalText_single);
+            Assert.IsNotEmpty(naturalText_single);
             Assert.AreEqual("2 maanden", naturalText_plural);
             Assert.AreEqual("1 maand", naturalText_single);
         }
@@ -133,8 +143,10 @@ namespace DateTimeExtensions.Tests
             var naturalText_plural = fromTime.ToExactNaturalText(toTime_plural, this.culture);
             var naturalText_single = fromTime.ToExactNaturalText(toTime_single, this.culture);
 
-            Assert.IsNotNullOrEmpty(naturalText_plural);
-            Assert.IsNotNullOrEmpty(naturalText_single);
+            Assert.IsNotNull(naturalText_plural);
+            Assert.IsNotEmpty(naturalText_plural);
+            Assert.IsNotNull(naturalText_single);
+            Assert.IsNotEmpty(naturalText_single);
             Assert.AreEqual("2 dagen, 2 uur, 2 minuten, 2 seconden", naturalText_plural);
             Assert.AreEqual("1 dag, 1 uur, 1 minuut, 1 seconde", naturalText_single);
         }
