+++ tests/DateTimeExtensions.Tests/project.json
@@ -1,0 +1,16 @@
+{
+  "version": "1.0.0-*",
+
+  "dependencies": {
+    "DateTimeExtensions": "1.0.0-*",
+    "NETStandard.Library": "1.6.0",
+    "Nunit": "3.6.0",
+    "NSubstitute": "2.0.0-beta"
+  },
+
+  "frameworks": {
+    "netstandard1.6": {
+      "imports": "dnxcore50"
+    }
+  }
+}
